Began rough draft of "User Profile"
Made a rough draft of the User Profile.
Let me know what you think and if you think any big changes need to happen.

Also fixed a few bugs when only 1 or 2 items were in the grid, causing the items to not start at the top of the container and instead float around the bottom/middle.